FINAL VERSION OF THE GAME

+++ BeerGame/src/Game.java
@@ -1,14 +1,18 @@
 import Field.Field;
+import Objects.Enemies.BadBeer;
+import Objects.Enemies.MC;
+import Objects.GoodBeer;
 import Objects.Movable;
 import Objects.ObjectFactory;
 import Objects.Player;
 import org.academiadecodigo.simplegraphics.graphics.Color;
 import org.academiadecodigo.simplegraphics.graphics.Text;
+import org.academiadecodigo.simplegraphics.pictures.Picture;
 
 public class Game {
 
     private Player player;
-    private final int DELAY = 2500;
+    private final int DELAY = 200;
     private ObjectList movables = new ObjectList();
     private CollisionDetector collisionDetector = new CollisionDetector();
 
@@ -23,22 +27,25 @@ public class Game {
 
         init();
 
-        Text score = new Text(50, 30, "Score: " + player.getScore());
-        score.grow(20, 20);
-        score.setColor(Color.GREEN);
+        Text score = new Text(50, 20, "Score: " + player.getScore());
+        score.grow(20, 10);
+        score.setColor(Color.WHITE);
         score.draw();
 
-        Text livers = new Text(700, 30, "Livers: " + player.getLivers());
-        livers.grow(20, 20);
-        livers.setColor(Color.PINK);
-        livers.draw();
+        Picture liver1 = new Picture(670, 10, "liver.png");
+        Picture liver2 = new Picture(710, 10, "liver.png");
+        Picture liver3 = new Picture(750, 10, "liver.png");
+        liver1.draw();
+        liver2.draw();
+        liver3.draw();
+
 
         while (!player.getDead()) {
 
 
-            try{
-                Thread.sleep((int) (Math.random() * DELAY + 100));
-            } catch (Exception e){
+            try {
+                Thread.sleep(DELAY);
+            } catch (Exception e) {
                 System.out.println("error");
             }
 
@@ -51,16 +58,50 @@ public class Game {
             for (Movable movable : movables) {
                 movable.move();
                 collisionDetector.check(player, movable);
-
                 score.setText("Score: " + player.getScore());
-                livers.setText("Livers: " + player.getLivers());
-
-                if(movable.getPosition().getRow() == Field.getRows()-1){
-                    movable.getPicture().delete();
+                if (player.getLivers() == 2) {
+                    liver3.delete();
                 }
+                if (player.getLivers() == 1) {
+                    liver2.delete();
+                }
+                if (player.getLivers() == 0) {
+                    liver1.delete();
+                }
+
+
+                if (movable.getPosition().getRow() == Field.getRows() - 1 && movable instanceof MC) {
+                    movable.getPicture().delete();
+                    ((MC) movable).isCollided(true);
+                }
+                if (movable.getPosition().getRow() == Field.getRows() - 1 && movable instanceof BadBeer) {
+                    movable.getPicture().delete();
+                    ((BadBeer) movable).isCollided(true);
+                }
+
+                if (movable.getPosition().getRow() == Field.getRows() - 1 && movable instanceof GoodBeer) {
+                    movable.getPicture().delete();
+                    ((GoodBeer) movable).isCollided(true);
+                }
+
             }
+
         }
+
+        player.getPicture().delete();
+        for (Movable movable : movables) {
+            movable.getPicture().delete();
+        }
+        Text gameOver = new Text(350, 250, "GAME OVER!");
+        gameOver.setColor(Color.RED);
+        gameOver.grow(200, 150);
+        gameOver.draw();
+        score.grow(20, 10);
+        score.translate(465, 350);
+
     }
 
 }
+
+
